+++ examples/advection_diffusion/vanilla_sdc.cpp
@@ -12,7 +12,7 @@
 #include <pfasst.hpp>
 #include <pfasst/logging.hpp>
 #include <pfasst/config.hpp>
-#include <pfasst/sdc.hpp>
+#include <pfasst/controller/sdc.hpp>
 #include <pfasst/encap/vector.hpp>
 
 #include "advection_diffusion_sweeper.hpp"
