+++ examples/boris/boris_sdc.cpp
@@ -3,7 +3,7 @@
 #include <pfasst.hpp>
 #include <pfasst/config.hpp>
 #include <pfasst/logging.hpp>
-#include <pfasst/sdc.hpp>
+#include <pfasst/controller/sdc.hpp>
 
 #include "particle.hpp"
 #include "particle_cloud.hpp"
