+++ examples/vanderpol/vdp_sdc.cpp
@@ -5,7 +5,7 @@
 #include <pfasst.hpp>
 #include <pfasst/config.hpp>
 #include <pfasst/logging.hpp>
-#include <pfasst/sdc.hpp>
+#include <pfasst/controller/sdc.hpp>
 #include <pfasst/encap/vector.hpp>
 
 #include "vdp_sweeper.hpp"
